refactored the tests about storage in subfiles
